rename gpjax.utils -> gpjax.typing
