+++ gpjax/gps.py
@@ -20,7 +20,7 @@ from beartype.typing import Any, Callable, Dict, Optional
 import jax.numpy as jnp
 from jax.random import PRNGKey, normal
 from jaxtyping import Array, Float
-from gpjax.utils import KeyArray
+from gpjax.typing import KeyArray
 
 from simple_pytree import static_field
