+++ gpjax/linops/linear_operator.py
@@ -21,7 +21,7 @@ from beartype.typing import Any, Generic, Iterable, Mapping, Tuple, Type, TypeVa
 import jax.numpy as jnp
 from jaxtyping import Array, Float
 from simple_pytree import Pytree, static_field
-from gpjax.utils import ScalarFloat
+from gpjax.typing import ScalarFloat
 
 # Generic type.
 T = TypeVar("T")
